test(topology): split randomized and unit tests to avoid duplicates
Having both test types in one class triggered our duplicated test
detection. Also, it's nicer to use the `RandomizedPropertyTest` suffix
to ensure that this test runs on its own job in CI.
